requirements.txt
streamlit
openai
feedparser
python-dotenv
beautifulsoup4
requests